distinguish server and local deployment

+++ app.py
@@ -4,8 +4,6 @@ from pytube import YouTube, StreamQuery
 
 import base64
 import os
-#import requests
-#import urllib.request
 
 # https://www.youtube.com/watch?v=Ch5VhJzaoaI&t=90s
 
@@ -21,17 +19,17 @@ def download_file(stream, fmt):
     else:
         title = stream.title + '.'+ stream_final.subtype
 
-    #stream.download(output_path='./', filename=title)
     stream.download(filename=title)
-    with open(title, 'rb') as f:
-        bytes = f.read()
-        b64 = base64.b64encode(bytes).decode()
-        href = f'<a href="data:file/zip;base64,{b64}" download=\'{title}1\'>\
-            Here is your link \
-        </a>'
-        st.markdown(href, unsafe_allow_html=True)
-        #response = requests.get(href)
-        #urllib.request.urlretrieve(href)
+    
+    if not os.environ['DESKTOP_SESSION']: #and os.environ('HOSTNAME')=='streamlit':
+    
+        with open(title, 'rb') as f:
+            bytes = f.read()
+            b64 = base64.b64encode(bytes).decode()
+            href = f'<a href="data:file/zip;base64,{b64}" download=\'{title}\'>\
+                Here is your link \
+            </a>'
+            st.markdown(href, unsafe_allow_html=True)
 
         os.remove(title)
 
@@ -69,7 +67,6 @@ st.set_page_config(page_title=" Youtube downloader", layout="wide")
 with st.sidebar:
 
     st.title("Youtube download app")
-    st.write(os.environ)
 
     url = st.text_input("Insert your link here", key="url")
 
@@ -101,7 +98,10 @@ with st.sidebar:
         # === Download block === #
         if stream_quality is not None:
             stream_final = stream_quality[0]
-            download = st.button("Download file", key='download')
+            if os.environ['DESKTOP_SESSION']:
+                download = st.button("Download file", key='download')
+            else:
+                download = st.button("Get download link", key='download')
 
             if download:
                 download_file(stream_final, fmt)
